Refactor api/mod.rs
